♻️ [Refactor]: mypage api endpoint 파일 변경
